updae onedevice login

+++ app/Http/Controllers/Api/Customer/CustomerApiController.php
@@ -24,7 +24,7 @@ class CustomerApiController extends Controller
         $customers=Customer::all();
         return response()->json($customers);
     }
-    
+
     /**
     * Show the form for creating a new resource.
     *
@@ -34,7 +34,7 @@ class CustomerApiController extends Controller
     {
         //
     }
-    
+
     /**
     * Store a newly created resource in storage.
     *
@@ -46,11 +46,11 @@ class CustomerApiController extends Controller
         $customer_phone=$request['customer_phone'];
         $fcm_token=$request['fcm_token'];
         $customer=Customer::where('customer_phone','=',$customer_phone)->first();
-        
+
         if($customer!=null){
             $customer->fcm_token=$fcm_token;
             $customer->update();
-            
+
             return response()->json(['success'=>true,'is_old'=>true,'message' => 'this is customer already exit','data'=>$customer]);
         }else{
             $customers=new Customer();
@@ -60,22 +60,52 @@ class CustomerApiController extends Controller
             $customers->save();
             return response()->json(['success'=>true,'is_old'=>false,'message' => 'this is customer create','data'=>$customers]);
         }
-        
+
     }
-    
+
     public function login_version_one(Request $request)
     {
+        $device_id = $request->header('device_id');
         $customer_phone=$request['customer_phone'];
         $fcm_token=$request['fcm_token'];
         $os_type=(int)$request['os_type'];
-        
+
         $customer=Customer::where('customer_phone','=',$customer_phone)->first();
-        
+
         if($customer!=null){
+
+            if($customer->device_id != $device_id){
+                $title="Another Device Login";
+                $messages="Your account have been login from another device";
+                $message = strip_tags($messages);
+                $path_to_fcm = 'https://fcm.googleapis.com/fcm/send';
+                $server_key = 'AAAAHUFURUE:APA91bFEvfAjoz58_u5Ns5l-y48QA9SgjICPzChgqVEg_S_l7ftvXrmGQjsE46rzGRRDtvGMnfqCWkksUMu0lDwdfxeTIHZPRMsdzFmEZx_0LIrcJoaUC-CF43XCxbMs2IMEgJNJ9j7E';
+                $header = array('Authorization:key=' . $server_key, 'Content-Type:application/json');
+                //Customer
+                $fcm_token_noti=array();
+                array_push($fcm_token_noti, $customer->fcm_token);
+
+                $notification = array('title' => $title, 'body' => $message,'sound'=>'default');
+                $field=array('registration_ids'=>$fcm_token_noti,'notification'=>$notification,'data'=>['type'=>'another_login','title' => $title,'body' => $message]);
+
+                $playLoad = json_encode($field);
+                $curl_session = curl_init();
+                curl_setopt($curl_session, CURLOPT_URL, $path_to_fcm);
+                curl_setopt($curl_session, CURLOPT_POST, true);
+                curl_setopt($curl_session, CURLOPT_HTTPHEADER, $header);
+                curl_setopt($curl_session, CURLOPT_RETURNTRANSFER, true);
+                curl_setopt($curl_session, CURLOPT_SSL_VERIFYPEER, false);
+                curl_setopt($curl_session, CURLOPT_IPRESOLVE, CURL_IPRESOLVE_V4);
+                curl_setopt($curl_session, CURLOPT_POSTFIELDS, $playLoad);
+                $result = curl_exec($curl_session);
+                curl_close($curl_session);
+            }
+
             $customer->fcm_token=$fcm_token;
+            $customer->device_id=$device_id;
             $customer->os_type=$os_type;
             $customer->update();
-            
+
             $check=ActiveCustomer::where('customer_id',$customer->customer_id)->whereDate('created_at',date('Y-m-d'))->first();
             if(empty($check)){
                 ActiveCustomer::create([
@@ -90,19 +120,19 @@ class CustomerApiController extends Controller
             $customers->image=null;
             $customers->os_type=$os_type;
             $customers->save();
-            
+
             ActiveCustomer::create([
                 "customer_id"=>$customers->customer_id,
             ]);
             return response()->json(['success'=>true,'is_old'=>false,'message' => 'this is customer create','data'=>$customers]);
         }
-        
+
     }
-    
+
     public function otp_send(Request $request)
     {
         $customer_phone=$request['customer_phone'];
-        
+
         $customer=Customer::where('customer_phone','=',$customer_phone)->first();
         $otp = sprintf("%06d", mt_rand(1, 999999));
         if($customer!=null){
@@ -114,14 +144,14 @@ class CustomerApiController extends Controller
             $response = $client->post($url,[
                 'headers' => ['Content-type' => 'application/json',
                 'Authorization' => 'Bearer ' .$token],
-                
+
                 'json' => [
                     "to"=>$customer_phone,
                     "message"=>$customer->otp." is your verification code for fatty application login",
                     "sender"=>"Fatty"
                 ],
             ]);
-            
+
             $result = json_decode($response->getBody());
             return response()->json(['success'=>true,'message' => 'Success OTP','data'=>$result]);
         }else{
@@ -129,38 +159,38 @@ class CustomerApiController extends Controller
             $customers->customer_phone=$customer_phone;
             $customers->otp=$otp;
             $customers->save();
-            
+
             ActiveCustomer::create([
                 "customer_id"=>$customers->customer_id,
             ]);
-            
+
             $client = new Client();
             $token = 'DPEL6xrzM-qqBqeVqmrOGP6jedLVpD5Z2r0D3Cun6IOCg3aFZVBqAYYJh4WA-CaF';
             $url = "https://smspoh.com/api/v2/send";
             $response = $client->post($url,[
                 'headers' => ['Content-type' => 'application/json',
                 'Authorization' => 'Bearer ' .$token],
-                
+
                 'json' => [
                     "to"=>$customer_phone,
                     "message"=>$customers->otp." is your verification code for fatty application login",
                     "sender"=>"Fatty"
                 ],
             ]);
-            
+
             $result = json_decode($response->getBody());
             return response()->json(['success'=>true,'message' => 'Success OTP','data'=>$result]);
         }
-        
+
     }
 
     public function otp_check(Request $request)
     {
         $customer_phone= $request['customer_phone'];
         $otp =  $request['otp'];
-        
+
         $customer=Customer::where('customer_phone','=',$customer_phone)->first();
-      
+
         if($customer!=null){
             if ($customer->otp == $otp) {
                 return response()->json(['success'=>true,'message' => 'OTP Check Success']);
@@ -170,9 +200,9 @@ class CustomerApiController extends Controller
         }else{
             return response()->json(['success'=>false,'message' => 'Customer phone not found']);
         }
-        
+
     }
-    
+
     /**
     * Display the specified resource.
     *
@@ -183,7 +213,7 @@ class CustomerApiController extends Controller
     {
         //
     }
-    
+
     /**
     * Show the form for editing the specified resource.
     *
@@ -194,7 +224,7 @@ class CustomerApiController extends Controller
     {
         //
     }
-    
+
     /**
     * Update the specified resource in storage.
     *
@@ -203,7 +233,8 @@ class CustomerApiController extends Controller
     * @return \Illuminate\Http\Response
     */
     public function update(Request $request)
-    {   
+    {
+        $device_id = $request->header('device_id');
         $id=$request['customer_id'];
         $customer_name=$request['customer_name'];
         $customer_phone=$request['customer_phone'];
@@ -211,28 +242,34 @@ class CustomerApiController extends Controller
         $image=$request['image'];
         $base_code_of_image=base64_decode($image);
         $imagename=$request['customer_phone'].time().'.jpg';
-        
+
         $customers=Customer::where('customer_id','=',$id)->first();
         if($customers){
-            $customers->customer_name = $customer_name;
-            $customers->customer_phone = $customer_phone;
-            $customers->fcm_token = $fcm_token;
-            
-            if($image){
-                if(!empty($customers->image)){
-                    Storage::disk('CustomersImages')->delete($customers->image);
+            $phone_check=Customer::where('customer_phone',$customer_phone)->where('customer_id','!=',$id)->first();
+            if($phone_check){
+                return response()->json(["success"=>false,"message"=>"the customer phone exits another customer"]);
+            }else{
+                $customers->customer_name = $customer_name;
+                $customers->customer_phone = $customer_phone;
+                $customers->fcm_token = $fcm_token;
+                $customers->device_id = $device_id;
+
+                if($image){
+                    if(!empty($customers->image)){
+                        Storage::disk('CustomersImages')->delete($customers->image);
+                    }
+                    $customers->image=$imagename;
+                    file_put_contents('uploads/customer/'.$imagename,$base_code_of_image);
                 }
-                $customers->image=$imagename;
-                file_put_contents('uploads/customer/'.$imagename,$base_code_of_image);
+
+                $customers->update();
+                return response()->json(['success'=>true,'message' => 'the customer have been updated','data'=>$customers]);
             }
-            
-            $customers->update();
-            return response()->json(['success'=>true,'message' => 'the customer have been updated','data'=>$customers]);
         }else{
             return response()->json(["success"=>false,"message"=>"the customer cannot update because customer cannot found in this data"]);
         }
     }
-    
+
     /**
     * Remove the specified resource from storage.
     *
@@ -249,9 +286,9 @@ class CustomerApiController extends Controller
         }else{
             return response()->json(['success'=>false,'message'=>'Error! customer_id not found']);
         }
-        
+
     }
-    
+
     public function location(Request $request)
     {
         $customer_id=$request['customer_id'];
@@ -262,7 +299,7 @@ class CustomerApiController extends Controller
             $customers->latitude=$latitude;
             $customers->longitude=$longitude;
             $customers->update();
-            
+
             $check=ActiveCustomer::where('customer_id',$customer_id)->whereDate('created_at',date('Y-m-d'))->first();
             if(empty($check)){
                 ActiveCustomer::create([
